+++ zagradio_me_front/lib/main.dart
@@ -3,8 +3,6 @@ import 'package:zagradio_me/screens/login_screen.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 
 Future<void> main() async {
-  
-  await dotenv.load(fileName: ".env"); 
 
   runApp(const MyApp());
 }
