+++ zagradio_me_front/pubspec.yaml
@@ -30,10 +30,11 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  google_maps_flutter: ^2.0.6  # For Google Maps integration
+  google_maps_flutter: ^2.9.0  # For Google Maps integration
+  google_maps_flutter_web: ^0.5.10
   geolocator: ^9.0.0            # For location services
   image_picker: ^0.8.4           # For image picking from camera and gallery
-  permission_handler: ^10.0.0     # For handling permissions 
+  permission_handler: ^10.4.0     # For handling permissions 
   intl: ^0.17.0    
   flutter_dotenv: ^5.1.0
